القبولُ الحيُّ كشف أنّ سكربت المصالحة لم يكن قابلاً للتنفيذ أصلاً
JSONB-PARAMETER-TYPE-UNDETERMINABLE-ON-LIVE-PG-01

CI كذّبت شاهدي الحيّ — وهذا نجاحُه لا فشلُه: لم يُتخطَّ، بل عمل وكشف.
ثلاثةُ أعطالٍ لي في المُهيّئ، ورابعٌ ليس لي وهو الأخطر.

أعطالي: `iot_devices` يشترط `name` و`type` NOT NULL (و`type` تحت CHECK)،
و`field_id` مفتاحٌ أجنبيٌّ إلى `fields` فلا يُربَط جهازٌ بحقلٍ غيرِ موجود؛
و`telemetry_id` مفتاحٌ أوّليٌّ **عابرٌ للمستأجرين** ففرضُ 10/11 يتصادم في قاعدة
CI المشتركة — صار يُولَّد بـRETURNING. ولم يمسكها قياسي المحلّيّ لأنّ هذه
الحاوية بلا PostGIS فتعذّر تطبيق v9/v24؛ فبنيتُ هذه المرّة مخطّطاً أدنى
حقيقيّاً وشغّلتُ الشاهدَ الحيّ عليه **قبل** الدفع.

والرابع: jsonb_build_object تأخذ anyelement، فوسيطٌ عارٍ داخلها بلا نوعٍ
قابلٍ للاستنتاج وPostgreSQL **يرفض PREPARE** الجملة
(IndeterminateDatatypeError). العطلُ في **الذراعَين** — soil_readings
وdevice_telemetry — و**قائمٌ على main منذ كُتِب السكربت**: مقيسٌ بتشغيل شكل
origin/main نفسِه على PostgreSQL 16 محلّيّاً، فيسقط. وسقوطُه كاملٌ لا جزئيّ:
لا صفَّ واحداً يُدرَج.

ولِمَ لم يكتشفه أحد: السجلُّ يقول عن السكربت إنّه «أداةُ مشغّلٍ لا يستدعيها
عاملٌ ولا مجدوِلٌ ولا سيرُ عمل» — وما لا يُشغَّل لا يفشل. والشاهدُ الساكن لا
يراه **بنيويّاً**: بديلُ الاتّصال لا يُنفِّذ PREPARE ولا يعرف أنواع PostgreSQL.

فشريحةُ الاستئناف لم تُصلح نموذجَ الاستئناف فحسب — كشفت أنّ المسار كلَّه لم
يكن قابلاً للتنفيذ. الإصلاحُ ::bigint على الوسيطَين، ومُكذَّبٌ بالزرع: نزعُ
التحويل يُسقط شاهدَين من ثلاثة على قاعدةٍ حيّة.

والدرسُ الذي أحمله: «مُكذَّبٌ بالزرع» في بديل اتّصالٍ يقيس منطقَ الدالّة لا
عقدَ الخادم. وقياسي لعقد v231 وحدَه لم يكن كافياً.

المقيس على هذا الرأس:
  الشاهدُ الحيُّ على PostgreSQL 16 حقيقيّ      3/3
  الشاهدُ الساكن                               15/15
  الطفرتان السلوكيّتان                         guard_mutation_guard_ok
  db_writer_ownership_guard                    ok baseline=74
  test_p0_db_ownership_guard                   2/2
  gate01_frozen_path_guard (فرقُ الطلب كاملاً)  ok
  regenerate ثمّ verify --check                rc=0 · 5923 بصمة · الحزمةُ آخِراً
  preflight --fast                             إخفاقات=0 · متخطّاة=0

حدُّ صدق: الإصلاحُ يجعل الجملة قابلةً للتحضير والتنفيذ، ولا يدّعي صحّةَ
المصالحة دلاليّاً ولا أنّها شُغِّلت على بيانات إنتاج ولا خلوَّ بقيّة السكربت
من أصنافٍ أخرى لا يراها إلّا خادمٌ حقيقيّ.

runtime_verified=0؛ production_certified=false.

+++ scripts/soil/reconcile_historical.py
@@ -37,6 +37,16 @@ SOIL_READING_PROPERTIES = (
     ("potassium", "potassium_mg_kg", "mg/kg"),
     ("organic_matter", "organic_matter_pct", "%"),
 )
+# JSONB-PARAMETER-TYPE-UNDETERMINABLE-ON-LIVE-PG-01. `jsonb_build_object` takes
+# `anyelement`, so a bare `$n` inside it has no inferable type and PostgreSQL refuses
+# to PREPARE the statement: `IndeterminateDatatypeError: could not determine data type
+# of parameter`. Both arms carried this since the script was written, and nothing
+# caught it because nothing runs this script against a live database — the gap record
+# says so in as many words: no worker, no scheduler, no workflow calls it. The offline
+# witness could not catch it either: its connection double never PREPAREs anything.
+# Found by the live acceptance added for RECONCILIATION-CURSOR-SKIPS-ROWS-THAT-BECOME-
+# ELIGIBLE-01, on the first run that reached a real server. The `::bigint` cast is the
+# whole fix; the legacy ids are BIGSERIAL in both source tables.
 TELEMETRY_MAP = {
     "soil_moisture": ("soil_moisture", "%"),
     "soil_temperature": ("soil_temperature", "degC"),
@@ -148,7 +158,7 @@ async def reconcile_soil_readings(conn, tenant_id: str, batch: int) -> Stats:
                        confidence, idempotency_key, provenance)
                    VALUES ($1,'soil-observation.v1',$2::uuid,$3,$4,to_jsonb($5::numeric),$6,
                            0,$7,$8,$8,'sensor',$9,'legacy_soil_readings_backfill',$10,'[]'::jsonb,
-                           0.65,$11,jsonb_build_object('legacy_table','soil_readings','legacy_id',$12))
+                           0.65,$11,jsonb_build_object('legacy_table','soil_readings','legacy_id',$12::bigint))
                    ON CONFLICT (tenant_id,idempotency_key) DO NOTHING""",
                 f"obs_legacy_sr_{row['id']}_{prop}",
                 tenant_id,
@@ -223,7 +233,7 @@ async def _process_telemetry_row(
                procedure_id,quality_status,quality_flags,confidence,idempotency_key,provenance)
            VALUES ($1,'soil-observation.v1',$2::uuid,$3,$4,to_jsonb($5::numeric),$6,0,30,$7,$8,
                    'sensor',$9,'device_telemetry_backfill','suspect','["depth_unknown","calibration_unknown"]'::jsonb,
-                   0.60,$10,jsonb_build_object('legacy_table','device_telemetry','legacy_id',$11))
+                   0.60,$10,jsonb_build_object('legacy_table','device_telemetry','legacy_id',$11::bigint))
            ON CONFLICT(tenant_id,idempotency_key) DO NOTHING""",
         f"obs_legacy_dt_{row['telemetry_id']}_{prop}",
         tenant_id,
